feat(sourcegen): split logic and add references support
