+++ lua/omnisharp_extended/log.lua
@@ -9,7 +9,7 @@
 -- User configuration section
 local default_config = {
   -- Name of the plugin. Prepended to log messages
-  plugin = 'omnisharp-extended-lsp.nvim',
+  plugin = "omnisharp-extended-lsp.nvim",
 
   -- Should print the output to neovim while running
   use_console = true,
@@ -25,12 +25,12 @@ local default_config = {
 
   -- Level configuration
   modes = {
-    { name = "trace", hl = "Comment", },
-    { name = "debug", hl = "Comment", },
-    { name = "info",  hl = "None", },
-    { name = "warn",  hl = "WarningMsg", },
-    { name = "error", hl = "ErrorMsg", },
-    { name = "fatal", hl = "ErrorMsg", },
+    { name = "trace", hl = "Comment" },
+    { name = "debug", hl = "Comment" },
+    { name = "info", hl = "None" },
+    { name = "warn", hl = "WarningMsg" },
+    { name = "error", hl = "ErrorMsg" },
+    { name = "fatal", hl = "ErrorMsg" },
   },
 
   -- Can limit the number of decimals displayed for floats
@@ -45,7 +45,7 @@ local unpack = unpack or table.unpack
 log.new = function(config, standalone)
   config = vim.tbl_deep_extend("force", default_config, config)
 
-  local outfile = string.format('%s/%s.log', vim.api.nvim_call_function('stdpath', {'data'}), config.plugin)
+  local outfile = string.format("%s/%s.log", vim.api.nvim_call_function("stdpath", { "data" }), config.plugin)
 
   local obj
   if standalone then
@@ -62,12 +62,12 @@ log.new = function(config, standalone)
   local round = function(x, increment)
     increment = increment or 1
     x = x / increment
-    return (x > 0 and math.floor(x + .5) or math.ceil(x - .5)) * increment
+    return (x > 0 and math.floor(x + 0.5) or math.ceil(x - 0.5)) * increment
   end
 
   local make_string = function(...)
     local t = {}
-    for i = 1, select('#', ...) do
+    for i = 1, select("#", ...) do
       local x = select(i, ...)
 
       if type(x) == "number" and config.float_precision then
@@ -83,7 +83,6 @@ log.new = function(config, standalone)
     return table.concat(t, " ")
   end
 
-
   local log_at_level = function(level, level_config, message_maker, ...)
     -- Return early if we're below the config.level
     if level < levels[config.level] then
@@ -97,13 +96,7 @@ log.new = function(config, standalone)
 
     -- Output to console
     if config.use_console then
-      local console_string = string.format(
-      "[%-6s%s] %s: %s",
-      nameupper,
-      os.date("%H:%M:%S"),
-      lineinfo,
-      msg
-      )
+      local console_string = string.format("[%-6s%s] %s: %s", nameupper, os.date("%H:%M:%S"), lineinfo, msg)
 
       if config.highlights and level_config.hl then
         vim.cmd(string.format("echohl %s", level_config.hl))
@@ -122,8 +115,7 @@ log.new = function(config, standalone)
     -- Output to log file
     if config.use_file then
       local fp = io.open(outfile, "a")
-      local str = string.format("[%-6s%s] %s: %s\n",
-      nameupper, os.date(), lineinfo, msg)
+      local str = string.format("[%-6s%s] %s: %s\n", nameupper, os.date(), lineinfo, msg)
       fp:write(str)
       fp:close()
     end
@@ -134,9 +126,9 @@ log.new = function(config, standalone)
       return log_at_level(i, x, make_string, ...)
     end
 
-    obj[("fmt_%s" ):format(x.name)] = function()
+    obj[("fmt_%s"):format(x.name)] = function()
       return log_at_level(i, x, function(...)
-        local passed = {...}
+        local passed = { ... }
         local fmt = table.remove(passed, 1)
         local inspected = {}
         for _, v in ipairs(passed) do
